+++ src/application/Application.cpp
@@ -1,38 +1,20 @@
 #include "Application.h"
 #include "core/logging/Logger.h"
 
-void Application::Update() {
-    float deltaTime = time->GetDeltaTime();
-
-    OnUpdate(deltaTime);
-}
-
 Application::Application(int width, int height, const std::string &title)
-    : window(nullptr), input(nullptr), time(nullptr),
-      isRunning(false) {
-    window = std::make_unique<Window>(width, height, title);
+    : isRunning(false) {
+    moduleManager = std::make_unique<ModuleManager>();
 
-    Logger::Log(LogLevel::INFO, "Application created");
+    m_coreModule = moduleManager->RegisterModule<CoreModule>(width, height, title);
+    m_coreModule->Initialize();
+    if (!m_coreModule->IsInitialized()) Logger::Log(LogLevel::CRITICAL, "Core failed to initialize");
 }
-
 
 Application::~Application() {
     Shutdown();
 }
 
 bool Application::Initialize() {
-    if (!window->Initialize()) {
-        Logger::Log(LogLevel::ERROR, "Failed to initialize Window");
-        return false;
-    }
-
-    glfwSetWindowUserPointer(window->GetGLFWWindow(), this);
-
-    input = std::make_unique<Input>(window->GetGLFWWindow());
-    time = std::make_unique<Time>();
-
-    moduleManager = std::make_unique<ModuleManager>();
-
     Logger::AddSink(&console);
     // Logger::AddSink(&file);
 
@@ -46,14 +28,14 @@ bool Application::Initialize() {
 void Application::Run() {
     isRunning = true;
 
-    while (isRunning && !window->ShouldClose()) {
-        time->Update();
+    while (isRunning && !m_coreModule->GetWindow()->ShouldClose()) {
+        m_coreModule->GetTime()->Update();
 
         Update();
         OnRender();
 
-        window->SwapBuffers();
-        window->PollEvents();
+        m_coreModule->GetWindow()->SwapBuffers();
+        m_coreModule->GetWindow()->PollEvents();
     }
 }
 
@@ -70,18 +52,12 @@ void Application::Stop() {
     isRunning = false;
 }
 
-Window *Application::GetWindow() {
-    return window.get();
-}
-
-Input *Application::GetInput() {
-    return input.get();
-}
-
-Time *Application::GetTime() const {
-    return time.get();
-}
-
 ModuleManager *Application::GetModuleManager() {
     return moduleManager.get();
+}
+
+void Application::Update() {
+    float deltaTime = m_coreModule->GetTime()->GetDeltaTime();
+
+    OnUpdate(deltaTime);
 }