+++ src/core/EventBus.h
@@ -6,16 +6,24 @@
 #include <any>
 #include <functional>
 
+using Callback = std::function<void(const std::any &)>;
+using EventID = std::string;
+using SubscriberID = size_t;
+
 class EventBus {
-    using Callback = std::function<void(const std::any &)>;
-    std::unordered_map<std::string, std::vector<Callback> > m_Subscribers;
+    std::unordered_map<EventID, std::vector<std::pair<SubscriberID, Callback>>> m_Subscribers;
+    std::vector<std::pair<EventID, SubscriberID>> m_PendingUnsubscribes;
+
+    bool m_IsPublishing = false;
+    SubscriberID m_NextID = -1;
 
 public:
-    void RegisterEvent(const std::string &name);
+    void RegisterEvent(const EventID &name);
 
-    void Subscribe(const std::string &name, Callback callback);
+    SubscriberID Subscribe(const EventID &name, Callback callback);
+    void Unsubscribe(const EventID &name, SubscriberID id);
 
-    void Publish(const std::string &name, const std::any &payload = {});
+    void Publish(const EventID &name, const std::any &payload = {});
 };
 
 EventBus &GetEventBus();