+++ src/core/Time.h
@@ -8,7 +8,6 @@
 /// @date 2026-01-28
 
 class Time {
-private:
     float currentFrame;
     float lastFrame;
     float deltaTime;
@@ -17,18 +16,12 @@ private:
 public:
     Time();
 
-
     void Update();
 
-
     float GetDeltaTime() const;
-
     float GetTime() const;
-
     float GetTimeScale() const;
-
     void SetTimeScale(float scale);
-
 
     float GetFPS() const;
 };