+++ src/engine/Engine.cpp
@@ -1,28 +1,32 @@
 #include "Engine.h"
+
+#include <imgui.h>
+
 #include <ImGuizmo.h>
 #include <glm/glm.hpp>
 #include "core/Input.h"
 #include "core/CommandManager.h"
 #include "core/logging/Logger.h"
+#include "EngineCommandHandler.h"
 
 void Engine::FramebufferSizeCallback(GLFWwindow *window, int width, int height) {
     glViewport(0, 0, width, height);
     Application *app = static_cast<Application *>(glfwGetWindowUserPointer(window));
-    if (app && app->GetWindow())
-        app->GetWindow()->SetSize(width, height);
+    if (app && app->GetModuleManager()->GetModule<CoreModule>("Core")->GetWindow())
+        app->GetModuleManager()->GetModule<CoreModule>("Core")->GetWindow()->SetSize(width, height);
 }
 
 void Engine::MouseCallback(GLFWwindow *window, double xpos, double ypos) {
     Engine *engine = static_cast<Engine *>(glfwGetWindowUserPointer(window));
     if (!engine) return;
 
-    engine->GetInput()->UpdateMousePosition(xpos, ypos);
+    engine->mm->GetModule<CoreModule>("Core")->GetInput()->UpdateMousePosition(xpos, ypos);
 
     if (ImGuizmo::IsUsing() || ImGui::GetIO().WantCaptureMouse)
         return;
 
     if (engine->cameraControlEnabled && engine->mainCameraEntity != entt::null) {
-        glm::vec2 delta = engine->GetInput()->GetMouseDelta();
+        glm::vec2 delta = engine->mm->GetModule<CoreModule>("Core")->GetInput()->GetMouseDelta();
         auto &orientation = engine->ecsModule->GetECS()->GetComponent<CameraOrientationComponent>(
             engine->mainCameraEntity);
         auto &config = engine->ecsModule->GetECS()->GetComponent<CameraComponent>(engine->mainCameraEntity);
@@ -48,10 +52,10 @@ void Engine::SetCameraControlMode(bool enabled) {
     cameraControlEnabled = enabled;
 
     if (enabled) {
-        GetWindow()->SetCursorMode(GLFW_CURSOR_DISABLED);
+        mm->GetModule<CoreModule>("Core")->GetWindow()->SetCursorMode(GLFW_CURSOR_DISABLED);
         Logger::Log(LogLevel::INFO, "Camera control: ON");
     } else {
-        GetWindow()->SetCursorMode(GLFW_CURSOR_NORMAL);
+        mm->GetModule<CoreModule>("Core")->GetWindow()->SetCursorMode(GLFW_CURSOR_NORMAL);
         Logger::Log(LogLevel::INFO, "Camera control: OFF (UI mode)");
     }
 }
@@ -60,82 +64,65 @@ void Engine::OnInitialize() {
     Logger::Log(LogLevel::INFO, "Initializing WFE...");
     Logger::Log(LogLevel::INFO, "==================================");
 
-    GetWindow()->SetFramebufferSizeCallback(FramebufferSizeCallback);
-    GetWindow()->SetCursorPosCallback(MouseCallback);
-    GetWindow()->SetScrollCallback(Input::ScrollCallback);
-    GetWindow()->SetMouseButtonCallback(MouseButtonCallback);
-
     mm = GetModuleManager();
+
+    auto *core = mm->GetModule<CoreModule>("Core");
+    auto *window = core->GetWindow();
+    auto *glfwWin = window->GetGLFWWindow();
+
+    glfwSetWindowUserPointer(glfwWin, this);
+
+    mm->GetModule<CoreModule>("Core")->GetWindow()->SetFramebufferSizeCallback(FramebufferSizeCallback);
+    mm->GetModule<CoreModule>("Core")->GetWindow()->SetCursorPosCallback(MouseCallback);
+    mm->GetModule<CoreModule>("Core")->GetWindow()->SetScrollCallback(Input::ScrollCallback);
+    mm->GetModule<CoreModule>("Core")->GetWindow()->SetMouseButtonCallback(MouseButtonCallback);
 
     mm->RegisterModule<ECSModule>();
     ecsModule = mm->GetModule<ECSModule>("ECS");
     ecsModule->Initialize();
-    if (!ecsModule->IsInitialized()) {
-        Logger::Log(LogLevel::WARNING,
-                    "ECS failed to initialize");
-    }
+    if (!ecsModule->IsInitialized()) Logger::Log(LogLevel::CRITICAL, "ECS failed to initialize");
 
     mm->RegisterModule<ResourceModule>();
     resourceModule = mm->GetModule<ResourceModule>("Resource");
     resourceModule->Initialize();
-    if (!resourceModule->IsInitialized()) {
-        Logger::Log(LogLevel::WARNING,
-                    "RedsourceModule failed to initialize");
-    }
+    if (!resourceModule->IsInitialized()) Logger::Log(LogLevel::CRITICAL, "RedsourceModule failed to initialize");
 
-    mm->RegisterModule<RenderingModule>(
-        GetWindow()->GetGLFWWindow(),
-        ecsModule->GetECS(),
-        mm);
+    mm->RegisterModule<RenderingModule>(mm->GetModule<CoreModule>("Core")->GetWindow()->GetGLFWWindow(),
+        ecsModule->GetECS(), mm);
     renderingModule = mm->GetModule<RenderingModule>("Rendering");
     renderingModule->Initialize();
-    if (!renderingModule->IsInitialized()) {
-        Logger::Log(LogLevel::WARNING,
-                    "RenderingModule failed to initialize");
-    }
+    if (!renderingModule->IsInitialized()) Logger::Log(LogLevel::CRITICAL, "RenderingModule failed to initialize");
 
     mm->RegisterModule<SceneModule>(ecsModule->GetECS());
     sceneModule = mm->GetModule<SceneModule>("Scene");
     sceneModule->Initialize();
-    if (!sceneModule->IsInitialized()) {
-        Logger::Log(LogLevel::WARNING,
-                    "SceneModule failed to initialize");
-    }
+    if (!sceneModule->IsInitialized()) Logger::Log(LogLevel::CRITICAL, "SceneModule failed to initialize");
 
     mm->RegisterModule<PhysicsModule>(ecsModule->GetECS());
     m_physicsModule = mm->GetModule<PhysicsModule>("Physics");
     m_physicsModule->Initialize();
-    if (!m_physicsModule->IsInitialized()) {
-        Logger::Log(LogLevel::WARNING,
-                    "PhysicsModule failed to initialize");
-    }
+    if (!m_physicsModule->IsInitialized()) Logger::Log(LogLevel::CRITICAL, "PhysicsModule failed to initialize");
 
-    mm->RegisterModule<UIModule>(
-        ecsModule->GetECS(),
-        &mainCameraEntity,
-        sceneModule->GetSceneManager(),
-        mm,
-        GetWindow()->GetGLFWWindow()
+    m_scriptModule = mm->RegisterModule<ScriptModule>(mm);
+    m_scriptModule->Initialize();
+    if (!m_scriptModule->IsInitialized()) Logger::Log(LogLevel::CRITICAL, "ScriptModule failed to initialize");
+
+    mm->RegisterModule<UIModule>(ecsModule->GetECS(), &mainCameraEntity, sceneModule->GetSceneManager(), mm,
+        mm->GetModule<CoreModule>("Core")->GetWindow()->GetGLFWWindow()
     );
     uiModule = mm->GetModule<UIModule>("UI");
     uiModule->Initialize();
-    if (!uiModule->IsInitialized()) {
-        Logger::Log(LogLevel::WARNING,
-                    "UIModule failed to initialize");
-    }
+    if (!uiModule->IsInitialized()) Logger::Log(LogLevel::CRITICAL, "UIModule failed to initialize");
 
-    scriptSystem = std::make_unique<ScriptSystem>();
     inputControllerSystem = std::make_unique<InputControllerSystem>();
     physicsDebugSystem = std::make_unique<PhysicsDebugRenderSystem>();
-    audioSystem = std::make_unique<AudioSystem>();
-    if (!audioSystem->Init()) {
-        Logger::Log(LogLevel::WARNING, "AudioSystem failed to initialize");
-    }
 
-    InitializeAS();
+    // audioSystem = std::make_unique<AudioSystem>();
+    // if (!audioSystem->Init()) {
+    //     Logger::Log(LogLevel::CRITICAL, "AudioSystem failed to initialize");
+    // }
 
-    ech = std::make_unique<EditorCommandHandler>(mm);
-    ech->RegisterAllCommands();
+    EditorCommandHandler(mm).RegisterAllCommands();
     RegistraterCoreCommands();
 
     auto editorCam = ecsModule->GetECS()->CreateCamera("Main Camera", true, true);
@@ -151,23 +138,22 @@ void Engine::OnInitialize() {
 }
 
 void Engine::OnUpdate(float deltaTime) {
-    mm->UpdateAll(deltaTime);
-
     ProcessInput();
 
     bool allowCameraControl = cameraControlEnabled && ShouldAllowCameraControl();
     inputControllerSystem->Update(
         *ecsModule->GetECS(),
-        *GetInput(),
+        *mm->GetModule<CoreModule>("Core")->GetInput(),
         deltaTime,
         allowCameraControl
     );
 
     UpdateMainCamera();
 
-    scriptSystem->Update(*ecsModule->GetECS(), GetInput(), deltaTime);
-    if (audioSystem)
-        audioSystem->Update(ecsModule->GetECS());
+    // if (audioSystem)
+    //     audioSystem->Update(ecsModule->GetECS());
+
+    mm->UpdateAll(deltaTime);
 }
 
 void Engine::UpdateMainCamera() {
@@ -208,8 +194,8 @@ void Engine::OnRender() {
     renderer->Render(
         *ecs,
         camera,
-        GetWindow()->GetWidth(),
-        GetWindow()->GetHeight()
+        mm->GetModule<CoreModule>("Core")->GetWindow()->GetWidth(),
+        mm->GetModule<CoreModule>("Core")->GetWindow()->GetHeight()
     );
     renderer->EndFrame();
 
@@ -220,7 +206,8 @@ void Engine::OnRender() {
 
         glm::mat4 view = orientation.GetViewMatrix(transform.position);
         glm::mat4 projection = camComp.GetProjectionMatrix(
-            (float) GetWindow()->GetWidth() / (float) GetWindow()->GetHeight()
+            (float) mm->GetModule<CoreModule>("Core")->GetWindow()->GetWidth() /
+            (float) mm->GetModule<CoreModule>("Core")->GetWindow()->GetHeight()
         );
 
         physicsDebugSystem->Update(
@@ -230,9 +217,7 @@ void Engine::OnRender() {
             view,
             projection
         );
-    }
 
-    if (showUI) {
         renderingModule->GetRenderer()->GetIcon()->Update(
             *ecs,
             *resourceModule->GetShaderManager(),
@@ -240,7 +225,8 @@ void Engine::OnRender() {
             ecs->GetComponent<CameraOrientationComponent>(camera).GetViewMatrix(
                 ecs->GetComponent<TransformComponent>(camera).position),
             ecs->GetComponent<CameraComponent>(camera).GetProjectionMatrix(
-                (float) GetWindow()->GetWidth() / (float) GetWindow()->GetHeight())
+                (float) mm->GetModule<CoreModule>("Core")->GetWindow()->GetWidth() /
+                (float) mm->GetModule<CoreModule>("Core")->GetWindow()->GetHeight())
         );
 
         uiModule->GetImGuiManager()->BeginFrame();
@@ -253,8 +239,8 @@ void Engine::OnShutdown() {
     Logger::Log(LogLevel::INFO, "Shutting down engine...");
     Logger::Log(LogLevel::INFO, "==================================");
 
-    if (audioSystem)
-        audioSystem->Shutdown();
+    // if (audioSystem)
+    //     audioSystem->Shutdown();
 
     mm->ShutdownAll();
 
@@ -277,24 +263,13 @@ Engine::Engine(int w, int h, const std::string &title)
     : Application(w, h, title) {
 }
 
-void Engine::InitializeAS() {
-    Logger::Log(LogLevel::INFO, "Initializing AngelScript...");
-
-    try {
-        InitAS(ecsModule->GetECS(), GetInput(), audioSystem.get());
-        Logger::Log(LogLevel::INFO, "AngelScript initialized successfully");
-    } catch (const std::exception &e) {
-        Logger::Log(LogLevel::ERROR, "Failed to initialize AngelScript: " + std::string(e.what()));
-    }
-}
-
 void Engine::ProcessInput() {
-    if (GetInput()->IsKeyJustPressed(Key::KEY_F5)) {
+    if (mm->GetModule<CoreModule>("Core")->GetInput()->IsKeyJustPressed(Key::KEY_F5)) {
         resourceModule->GetShaderManager()->ReloadAll();
         Logger::Log(LogLevel::INFO, "Reloaded all shaders");
     }
 
-    if (GetInput()->IsKeyJustPressed(Key::KEY_F1))
+    if (mm->GetModule<CoreModule>("Core")->GetInput()->IsKeyJustPressed(Key::KEY_F1))
         showUI = !showUI;
 }