+++ src/engine/Engine.h
@@ -3,28 +3,20 @@
 #include <string>
 #include <memory>
 
-#include <imgui.h>
-#include <imgui_impl_glfw.h>
-#include <imgui_impl_opengl3.h>
 #include <entt/entt.hpp>
-#include <glad/glad.h>
 #include <GLFW/glfw3.h>
 
 #include "application/Application.h"
-#include "rendering/core/Framebuffer.h"
-#include "scripting/ASBindings.h"
 #include "ECS/systems/Systems.h"
 #include "ECS/components/Components.h"
-#include "EngineCommandHandler.h"
 #include "core/ModuleManager.h"
 #include "rendering/RenderingModule.h"
 #include "resource/ResourceModule.h"
 #include "UI/UIModule.h"
 #include "ECS/ECSModule.h"
 #include "scene/SceneModule.h"
-#include "core/EventBus.h"
-#include "UI/DebugOverlay.h"
 #include "physics/PhysicsModule.h"
+#include "scripting/ScriptModule.h"
 
 /// @file Engine.cppm
 /// @brief Engine class
@@ -34,19 +26,17 @@
  * 
  */
 class Engine : public Application {
-private:
-    std::unique_ptr<ScriptSystem> scriptSystem;
     std::unique_ptr<InputControllerSystem> inputControllerSystem;
     std::unique_ptr<PhysicsDebugRenderSystem> physicsDebugSystem;
-    std::unique_ptr<AudioSystem> audioSystem;
 
-    std::unique_ptr<EditorCommandHandler> ech;
+    //std::unique_ptr<AudioSystem> audioSystem;
 
     ModuleManager *mm;
     RenderingModule *renderingModule;
     ResourceModule *resourceModule;
     UIModule *uiModule;
     ECSModule *ecsModule;
+    ScriptModule *m_scriptModule;
     SceneModule *sceneModule;
     PhysicsModule *m_physicsModule;
 
@@ -81,7 +71,7 @@ protected:
     void OnRender() override;
 
     /**
-         * @brief Shutdown classes needed to shutdown by hand
+         * @brief Shutdown classes needed to shut down by hand
          */
     void OnShutdown() override;