+++ src/scripting/ASRegistration/ASRegisterAPI.h
@@ -7,12 +7,7 @@
 #include "ECS/World.h"
 #include "core/Input.h"
 
-#include "scripting/ASRegistration/ASRegisterCommands.h"
-#include "scripting/ASRegistration/ASRegisterECS.h"
-#include "scripting/ASRegistration/ASRegisterInput.h"
-#include "scripting/ASRegistration/ASRegisterMath.h"
-#include "scripting/ASRegistration/ASRegisterTypes.h"
-#include "scripting/ASRegistration/ASRegisterAudio.h"
+
 #include "ECS/systems/AudioSystem.h"
 
 class ASRegisterAPI {